requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.35.0
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.7.3
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
openai==1.34.0
httpx<0.28
faiss-cpu==1.8.0
python-dotenv==1.0.1
